requirements.txt
matplotlib==3.8.3
notebook==7.1.2
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
pytest==8.1.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
torch==2.2.1
torchmetrics==1.3.1
torchvision==0.17.1
tqdm==4.66.2
# this should install 4.9.0 at this moment (specifying version causes installation error on windows)
opencv-python
-e git+https://github.com/superleesa/TOENet.git@d9d4f3bb78f6d65ca816914959eb63130b688cf4#egg=toenet
pyyaml~=6.0.1
fire==0.6.0
flask==3.0.2
optuna== 3.6.1